Refactor the Database/AsyncDatabase god object into focused collaborators
