+++ test/async_/test_cypher.py
@@ -220,7 +220,7 @@ async def test_cypher_debug_log_masks_sensitive_params(caplog):
     original_debug = config.cypher_debug
     try:
         config.cypher_debug = True
-        with caplog.at_level(logging.DEBUG, logger="neomodel.async_.database"):
+        with caplog.at_level(logging.DEBUG, logger="neomodel.async_.query"):
             await adb.cypher_query("RETURN $password AS p", {"password": "s3cr3t"})
         # Only inspect neomodel's own log line: the neo4j driver has separate
         # protocol-level debug logging that echoes raw parameters and is outside
@@ -228,7 +228,7 @@ async def test_cypher_debug_log_masks_sensitive_params(caplog):
         logged = "\n".join(
             record.getMessage()
             for record in caplog.records
-            if record.name == "neomodel.async_.database"
+            if record.name == "neomodel.async_.query"
         )
         assert "s3cr3t" not in logged
         assert "******" in logged
@@ -248,14 +248,14 @@ async def test_cypher_debug_log_uses_custom_redaction_hook(caplog):
         config.cypher_log_redaction_hook = lambda params: {
             key: "<hidden>" for key in params
         }
-        with caplog.at_level(logging.DEBUG, logger="neomodel.async_.database"):
+        with caplog.at_level(logging.DEBUG, logger="neomodel.async_.query"):
             await adb.cypher_query("RETURN $email AS e", {"email": "user@example.com"})
         # Only inspect neomodel's own log line (see note above re: the driver's
         # separate protocol-level logging).
         logged = "\n".join(
             record.getMessage()
             for record in caplog.records
-            if record.name == "neomodel.async_.database"
+            if record.name == "neomodel.async_.query"
         )
         assert "user@example.com" not in logged
         assert "<hidden>" in logged
