+++ test/async_/test_database_management.py
@@ -98,7 +98,9 @@ async def test_change_password_is_not_injectable():
     malicious_user = "admin` SET PASSWORD 'pwned"
     malicious_password = "secret' SET ROLE admin //"
 
-    with patch.object(test_db, "cypher_query", new_callable=AsyncMock) as mock_cypher:
+    with patch.object(
+        test_db._query, "cypher_query", new_callable=AsyncMock
+    ) as mock_cypher:
         await test_db.change_neo4j_password(malicious_user, malicious_password)
 
     query, params = mock_cypher.call_args.args[:2]
