+++ test/sync_/test_core_additional.py
@@ -64,12 +64,12 @@ def test_clear_neo4j_database():
     """Test clear_neo4j_database method."""
     test_db = Database()
 
-    with patch.object(test_db, "cypher_query", new_callable=Mock) as mock_cypher:
+    with patch.object(test_db._query, "cypher_query", new_callable=Mock) as mock_cypher:
         with patch.object(
-            test_db, "drop_constraints", new_callable=Mock
+            test_db._schema, "drop_constraints", new_callable=Mock
         ) as mock_drop_constraints:
             with patch.object(
-                test_db, "drop_indexes", new_callable=Mock
+                test_db._schema, "drop_indexes", new_callable=Mock
             ) as mock_drop_indexes:
                 test_db.clear_neo4j_database(clear_constraints=True, clear_indexes=True)
 
@@ -88,7 +88,7 @@ def test_drop_constraints():
         {"name": "constraint2", "labelsOrTypes": ["Label2"], "properties": ["prop2"]},
     ]
 
-    with patch.object(test_db, "cypher_query", new_callable=Mock) as mock_cypher:
+    with patch.object(test_db._query, "cypher_query", new_callable=Mock) as mock_cypher:
         mock_cypher.return_value = (
             mock_results,
             ["name", "labelsOrTypes", "properties"],
@@ -110,10 +110,14 @@ def test_drop_indexes():
         {"name": "index2", "labelsOrTypes": ["Label2"], "properties": ["prop2"]},
     ]
 
-    with patch.object(test_db, "list_indexes", new_callable=Mock) as mock_list_indexes:
+    with patch.object(
+        test_db._schema, "list_indexes", new_callable=Mock
+    ) as mock_list_indexes:
         mock_list_indexes.return_value = mock_indexes
 
-        with patch.object(test_db, "cypher_query", new_callable=Mock) as mock_cypher:
+        with patch.object(
+            test_db._query, "cypher_query", new_callable=Mock
+        ) as mock_cypher:
             test_db.drop_indexes(quiet=False)
 
             # Should call DROP_INDEX_COMMAND for each index
@@ -126,10 +130,10 @@ def test_remove_all_labels():
     test_db = Database()
 
     with patch.object(
-        test_db, "drop_constraints", new_callable=Mock
+        test_db._schema, "drop_constraints", new_callable=Mock
     ) as mock_drop_constraints:
         with patch.object(
-            test_db, "drop_indexes", new_callable=Mock
+            test_db._schema, "drop_indexes", new_callable=Mock
         ) as mock_drop_indexes:
             with patch("sys.stdout") as mock_stdout:
                 test_db.remove_all_labels()
@@ -243,7 +247,9 @@ def test_cypher_query_client_error_generic():
     """Test cypher_query with generic ClientError."""
     test_db = Database()
 
-    with patch.object(test_db, "_run_cypher_query", new_callable=Mock) as mock_run:
+    with patch.object(
+        test_db._query, "_run_cypher_query", new_callable=Mock
+    ) as mock_run:
         client_error = ClientError("Neo.ClientError.Generic", "message")
         mock_run.side_effect = client_error
